Detecting & testing new D1 bindings
